move streetpass files into folder
