+++ include/streetpass/MessageInfo.hpp
